requirements.txt
absl-py==0.15.0
addict==2.4.0
appdirs==1.4.4
asgiref==3.5.2
astunparse==1.6.3
audioread==3.0.0
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
clang==5.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
djangorestframework==3.14.0
drf-yasg==1.21.4
fast_ctc_decode==0.3.2
filelock==3.8.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.1.0
huggingface-hub==0.11.1
idna==3.4
imagecodecs==2022.2.22
imageio==2.22.4
importlib-metadata==5.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jstyleson==0.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
midi @ git+https://github.com/vishnubob/python-midi@abb85028c97b433f74621be899a0b399cd100aaa
mido==1.2.10
mir-eval==0.7
networkx==2.8
nibabel==4.0.2
nltk==3.7
numba==0.56.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.6.0.66
openunmix==1.2.1
openvino==2022.2.0
openvino-dev==2022.2.0
openvino-telemetry==2022.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parasail==1.2.4
Pillow==9.3.0
pooch==1.6.0
pretty-midi==0.2.9
progress==1.6
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydicom==2.3.1
pydub==0.25.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
rawpy==0.17.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.97
Shapely==1.8.5.post1
six==1.15.0
soundfile==0.11.0
sqlparse==0.4.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.13.0
tqdm==4.64.1
transformers==4.25.1
typing-extensions==3.7.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
